added color picker slide in-out
