+++ src/web/src/components/mainPage/experimentsInspector/colorPicker/colorPicker.module.css
@@ -1,0 +1,82 @@
+/**
+ * Copyright (C) 2023 Hefestus
+ * 
+ * This file is part of Bolinho.
+ * 
+ * Bolinho is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Bolinho is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with Bolinho.  If not, see <http://www.gnu.org/licenses/>.
+ */
+@import url(../../../../variables.css);
+
+.color_picker_div {
+	flex: none;
+	z-index: 1;
+
+	position: relative;
+	height: fit-content;
+	animation: slide-out;
+	animation-fill-mode: forwards;
+	animation-duration: var(--animation_slow);
+	animation-timing-function: ease;
+}
+
+@keyframes slide-out {
+	0% {
+		opacity: 1;
+		transform: translateX(-100%) translateY(-100%);
+	}
+	100% {
+		opacity: 0;
+
+		transform: translateX(-100%) translateY(0%);
+	}
+}
+
+.color_picker_div_active {
+	animation: slide-in;
+	animation-fill-mode: forwards;
+	animation-duration: var(--animation_slow);
+	animation-timing-function: ease;
+}
+
+@keyframes slide-in {
+	0% {
+		opacity: 0;
+
+		transform: translateX(-100%) translateY(0%);
+	}
+	100% {
+		opacity: 1;
+
+		transform: translateX(-100%) translateY(-100%);
+	}
+}
+
+.color_picker {
+	/*
+	position: fixed;
+	z-index: 10;
+	transform: translateY(-50%);
+	float: right;
+    */
+}
+
+.color_picker_back_drop {
+	position: fixed;
+	left: 0;
+	top: 0;
+	right: 0;
+	bottom: 0;
+	background-color: rgba(0, 0, 0, 0.39);
+	pointer-events: all;
+}
