+++ src/web/src/components/mainPage/experimentsInspector/experimentButton/experimentButton.jsx
@@ -18,7 +18,6 @@
 import React, { useState, useEffect } from "react";
 import { getFormattedDate } from "../../../../helpers";
 import styleModule from "./experimentButton.module.css";
-import { toast } from "react-toastify";
 export default function ExperimentButton({
 	object,
 	materialName,
