+++ src/web/src/components/mainPage/experimentsInspector/experimentDescription/experimentDescription.module.css
@@ -1,0 +1,52 @@
+/**
+ * Copyright (C) 2023 Hefestus
+ * 
+ * This file is part of Bolinho.
+ * 
+ * Bolinho is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Bolinho is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with Bolinho.  If not, see <http://www.gnu.org/licenses/>.
+ */
+@import url(../../../../variables.css);
+
+.experiment_description {
+	background-color: var(--background_color);
+	padding: 2%;
+	width: 65%;
+	color: var(--font_color);
+	font-family: Roboto;
+	font-size: var(--font_s);
+	font-weight: 400;
+	text-align: left;
+	box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+	overflow-y: scroll;
+	padding-bottom: 0;
+
+	white-space: nowrap;
+	overflow-x: scroll;
+}
+.experiment_description h1 {
+	font-size: var(--font_xxs);
+	margin: 0;
+	margin-bottom: 0.5em;
+	margin-top: 0.5em;
+	text-align: center;
+}
+
+/* Im overwriting the p tag so it doesn't overflow */
+.experiment_description p {
+	font-size: var(--font_xxs);
+	margin: 0;
+	display: inline;
+	word-wrap: break-word;
+	white-space: break-spaces;
+}
