+++ src/web/src/components/mainPage/experimentsInspector/experimentsInspector.module.css
@@ -19,14 +19,19 @@
 
 @import url(../../../variables.css);
 
+.material_inspector_div {
+	height: stretch;
+	display: flex;
+}
 .material_inspector {
+	flex: none;
+	z-index: 2;
 	height: stretch;
 	width: stretch;
-	display: flex;
-	flex-direction: column;
 	border-radius: clamp(2px, 3vh, 20px);
 	overflow: hidden;
 	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+	position: relative;
 }
 
 .material_inspector_header {
@@ -106,13 +111,6 @@
 	mix-blend-mode: difference;
 }
 
-.material_inspector_header .color_picker {
-	position: fixed;
-	z-index: 10;
-	transform: translateY(-50%);
-	float: right;
-}
-
 .material_inspector_content {
 	height: calc(100% - clamp(20px, 5.5vh, 75px));
 	width: 100%;
@@ -138,37 +136,4 @@
 	list-style-type: none;
 	overflow-y: scroll;
 	overflow-x: hidden;
-}
-
-.material_inspector_description {
-	background-color: var(--background_color);
-	padding: 2%;
-	width: 65%;
-	color: var(--font_color);
-	font-family: Roboto;
-	font-size: var(--font_s);
-	font-weight: 400;
-	text-align: left;
-	box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
-	overflow-y: scroll;
-	padding-bottom: 0;
-
-	white-space: nowrap;
-	overflow-x: scroll;
-}
-.material_inspector_description h1 {
-	font-size: var(--font_xxs);
-	margin: 0;
-	margin-bottom: 0.5em;
-	margin-top: 0.5em;
-	text-align: center;
-}
-
-/* Im overwriting the p tag so it doesn't overflow */
-.material_inspector_description p {
-	font-size: var(--font_xxs);
-	margin: 0;
-	display: inline;
-	word-wrap: break-word;
-	white-space: break-spaces;
 }
